requirements.txt
requests
python-dotenv
pandas
pillow
numpy
pandas